CEDS-2091 Move NotificationFrontendModel to exchange package
