+++ test/unit/uk/gov/hmrc/exports/movements/models/notifications/NotificationFrontendModelSpec.scala
@@ -17,7 +17,7 @@
 package unit.uk.gov.hmrc.exports.movements.models.notifications
 
 import org.scalatest.{MustMatchers, WordSpec}
-import uk.gov.hmrc.exports.movements.models.notifications.NotificationFrontendModel
+import uk.gov.hmrc.exports.movements.models.notifications.exchange.NotificationFrontendModel
 import uk.gov.hmrc.exports.movements.models.notifications.standard.{Entry, UcrBlock}
 import utils.testdata.CommonTestData.ucr
 import utils.testdata.notifications.NotificationTestData._
